+++ src/main/java/controller/DeleteItem.java
@@ -15,13 +15,18 @@ import dto.FoodItem;
 public class DeleteItem extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		//Getting Id from the URL
-		int id=Integer.parseInt(req.getParameter("id"));
-		MyDao dao=new MyDao();
-		//Finding object because remove method accepts object
-		FoodItem item=dao.find(id);
-		dao.delete(item);
-		resp.getWriter().print("<h1 style='color:green'>Data Deleted Successfully</h1>");
-		req.getRequestDispatcher("viewmenu").include(req, resp);
+		if (req.getSession().getAttribute("admin") == null) {
+			resp.getWriter().print("<h1 style='color:red'>Invalid Session</h1>");
+			req.getRequestDispatcher("Login.html").include(req, resp);
+		} else {
+			// Getting Id from the URL
+			int id = Integer.parseInt(req.getParameter("id"));
+			MyDao dao = new MyDao();
+			// Finding object because remove method accepts object
+			FoodItem item = dao.find(id);
+			dao.delete(item);
+			resp.getWriter().print("<h1 style='color:green'>Data Deleted Successfully</h1>");
+			req.getRequestDispatcher("viewmenu").include(req, resp);
+		}
 	}
 }
